move test util libs
