+++ test/basic.js
@@ -1,6 +1,8 @@
 const tape = require('tape')
 const { Kappa } = require('..')
 const { runAll } = require('./lib/util')
+const createSimpleView = require('./lib/simple-view')
+const createSimpleSource = require('./lib/simple-source')
 
 tape('simple source', t => {
   const kappa = new Kappa()
@@ -245,81 +247,3 @@ tape('fetch version error', t => {
     t.end()
   })
 })
-
-function createSimpleView () {
-  let res = []
-  let clears = 0
-  const view = {
-    map (msgs, next) {
-      res = res.concat(msgs)
-      next()
-    },
-    reset (cb) {
-      clears = clears + 1
-      res = []
-      cb()
-    },
-    api: {
-      collect (kappa, cb) {
-        this.ready(() => cb(null, res))
-      },
-      clearedCount (kappa) {
-        return clears
-      }
-    }
-  }
-  return view
-}
-
-function createSimpleSource (opts = {}) {
-  const buf = []
-  const maxBatch = opts.maxBatch || 10
-  let flow = null
-  let state = 0
-  let error = null
-
-  const source = {
-    open (_flow, next) {
-      flow = _flow
-      next()
-    },
-    pull (next) {
-      if (error) return next({ error })
-      const max = buf.length
-      const end = Math.min(state + maxBatch, max)
-      const messages = buf.slice(state, end)
-      const lastState = state
-      next({
-        messages,
-        finished: end === max,
-        onindexed (cb) {
-          state = end
-          cb(null, {
-            totalBlocks: buf.length,
-            indexedBlocks: end,
-            prevIndexedBlocks: lastState
-          })
-        }
-      })
-    },
-    reset (next) {
-      state = 0
-      next()
-    },
-    get api () {
-      return {
-        push (kappa, value) {
-          if (!Array.isArray(value)) value = [value]
-          buf.push(...value)
-          if (flow) flow.update()
-        },
-        error (kappa, err) {
-          error = err
-          if (flow) flow.update()
-        }
-      }
-    }
-  }
-
-  return source
-}
